almost there

+++ ft_puthex_lo.c
@@ -12,7 +12,7 @@
 
 #include <unistd.h>
 
-static int	len_ft(unsigned int num)
+static int	len_ft(unsigned long num)
 {
 	int	len;
 
@@ -27,7 +27,7 @@ static int	len_ft(unsigned int num)
 	return (len);
 }
 
-int	ft_puthex_lo(unsigned int n)
+int	ft_puthex_lo(unsigned long n)
 {
 	char	*base;
 	int		len;
